Simplify caches
