+++ package.json
@@ -27,6 +27,7 @@
   },
   "dependencies": {
     "dot-json": "^1.2.2",
+    "fast-json-stable-stringify": "^2.1.0",
     "js-base64": "^3.7.1",
     "json2ts": "^0.0.7",
     "lodash": "^4.17.21",
