+++ template/src/util/ix.test.ts
@@ -8,8 +8,6 @@ import {
     asyncGeneratorFromPromise,
     concat,
     observableFromAsyncIterator,
-    cachePromiseProvider,
-    PROMISE_CACHE_CAPACITY,
 } from './ix'
 
 describe('observableFromAsyncIterator', () => {
@@ -82,83 +80,6 @@ describe('asyncGeneratorFromPromise', () => {
         const iterable = asyncGeneratorFromPromise(async (value: number) => Promise.resolve(value * 2))
 
         assert.deepStrictEqual(await gatherValues(iterable(24)), [48])
-    })
-})
-
-const textDocument1 = createStubTextDocument({
-    uri: 'https://sourcegraph.test/repo@rev/-/raw/foo.ts',
-    languageId: 'typescript',
-    text: undefined,
-})
-const textDocument2 = createStubTextDocument({
-    uri: 'https://sourcegraph.test/repo@rev/-/raw/bar.ts',
-    languageId: 'typescript',
-    text: undefined,
-})
-
-const textDocument3 = createStubTextDocument({
-    uri: 'https://sourcegraph.test/repo@rev/-/raw/baz.ts',
-    languageId: 'typescript',
-    text: undefined,
-})
-
-const position1 = new sourcegraph.Position(10, 1)
-const position2 = new sourcegraph.Position(10, 2)
-const position3 = new sourcegraph.Position(10, 3)
-
-describe('cachePromiseProvider', () => {
-    it('caches the result', async () => {
-        let calls = 0
-        let resolutions = 0
-
-        const cachedPromise = cachePromiseProvider(
-            (textDocument: sourcegraph.TextDocument, position: sourcegraph.Position) => {
-                calls++
-
-                return new Promise<number>(resolve => {
-                    resolutions++
-                    resolve(position.line + position.character * 2)
-                })
-            }
-        )
-
-        const promise = cachedPromise(textDocument1, position1)
-        assert.strictEqual(await promise, 12)
-        assert.strictEqual(await promise, 12)
-        assert.strictEqual(await cachedPromise(textDocument2, position2), 14)
-        assert.strictEqual(await cachedPromise(textDocument3, position3), 16)
-        assert.strictEqual(await cachedPromise(textDocument2, position2), 14)
-        assert.strictEqual(await cachedPromise(textDocument1, position1), 12)
-        assert.strictEqual(calls, 3)
-        assert.strictEqual(resolutions, 3)
-    })
-
-    it('is bounded', async () => {
-        let calls = 0
-
-        const cachedPromise = cachePromiseProvider(
-            (textDocument: sourcegraph.TextDocument, position: sourcegraph.Position) => {
-                calls++
-                return new Promise<number>(resolve => resolve(position.line + position.character * 2))
-            }
-        )
-
-        for (let index = 0; index < PROMISE_CACHE_CAPACITY * 2; index++) {
-            await cachedPromise(textDocument1, new sourcegraph.Position(index, 1))
-        }
-        assert.strictEqual(calls, PROMISE_CACHE_CAPACITY * 2)
-
-        // Ensure that the later half are still cached - no new calls
-        for (let index = 0; index < PROMISE_CACHE_CAPACITY; index++) {
-            await cachedPromise(textDocument1, new sourcegraph.Position(2 * PROMISE_CACHE_CAPACITY - index - 1, 1))
-        }
-        assert.strictEqual(calls, PROMISE_CACHE_CAPACITY * 2)
-
-        // Ensure that the first half is not cached - all new calls
-        for (let index = 0; index < PROMISE_CACHE_CAPACITY; index++) {
-            await cachedPromise(textDocument1, new sourcegraph.Position(index, 1))
-        }
-        assert.strictEqual(calls, PROMISE_CACHE_CAPACITY * 3)
     })
 })
